Fixed middleware issues
