+++ src/Core.Tests/Execution/QueryExecuterTests.txt
