+++ src/Core/Execution/QueryExecuterExtensions.cs
@@ -1,0 +1,15 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace HotChocolate.Execution
+{
+    public static class QueryExecuterExtensions
+    {
+        public static Task<IExecutionResult> ExecuteAsync(
+            this IQueryExecuter executer,
+            QueryRequest request)
+        {
+            return executer.ExecuteAsync(request, CancellationToken.None);
+        }
+    }
+}
